+++ src/app/app.module.ts
@@ -2,21 +2,21 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppComponent } from './app.component';
-import { FirstComponent } from './first/first.component';
-import { ProductComponent } from './feature/product/product.component';
-import { ProductListComponent } from './feature/product-list/product-list.component';
-import { CartListComponent } from './feature/cart-list/cart-list.component';
+import { CartListModule } from './feature/cart-list/cart-list.module';
+import { ProductListModule } from './feature/product-list/product-list.module';
+import { SharedModule } from './shared/shared.module';
+import { CartItemModule } from './feature/cart-item/cart-item.module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    FirstComponent,
-    ProductComponent,
-    ProductListComponent,
-    CartListComponent,
   ],
   imports: [
-    BrowserModule
+    BrowserModule,
+    CartListModule,
+    ProductListModule,
+    SharedModule,
+    CartItemModule,
   ],
   providers: [],
   bootstrap: [AppComponent]